refactors class and inspector in bdv source and converter service ui
